requirements.txt
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
cookiecutter==1.7.3
Django==4.0.1
django-browser-reload==1.3.0
django-tailwind==3.1.1
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==5.0.2
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2021.5
urllib3==1.26.8
djangorestframework==3.13.1
django-filter==21.1